requirements.txt
python-crfsuite==0.9.6
keras==2.3.1
tensorflow==2.3.0
numpy==1.18.5
tensorflow_hub==0.10.0
bert-for-tf2